make VoteLeaderboardSearch component, update Ballot to accomodate, begin making new route for leaderboard after vote has ended
