+++ src/views/Voting/VoteOverLeaderboard.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>
+    test
+  </div>
+</template>
+
+<script>
+export default {
+  // TODO: remain leaderboard to leaderbar, and rename this to leaderboard
+  name: "VoteOverLeaderboard"
+};
+</script>
+
+<style lang="scss" scoped></style>
